+++ packages/react-menu/src/MenuList.stories.tsx
@@ -298,8 +298,6 @@ export const RadioItemsControlled = () => {
 
 export default {
   // use the Components prefix and (react-menu) suffix to have the same naming convention as react-examples
-  title: 'Components/MenuList (react-menu)',
-  // Explicit id used in story URL
-  id: 'Components/MenuList',
+  title: 'Components/MenuList',
   component: MenuList,
 };