+++ packages/react-menu/src/components/Menu/useMenuPopup.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { getCode, keyboardKey, EnterKey, SpacebarKey } from '@fluentui/keyboard-key';
+import { getCode, keyboardKey } from '@fluentui/keyboard-key';
 import { MenuState } from './Menu.types';
 import { Portal } from '@fluentui/react-portal';
 import { isOutsideMenu } from '../../utils/index';
@@ -36,8 +36,6 @@ export const useMenuPopup = (state: UseMenuPopupState) => {
     inline,
   } = state;
 
-  const dismissedWithKeyboardRef = React.useRef(false);
-
   menuPopup.children = (Component, originalProps) => {
     const newProps = { role: 'presentation', ...originalProps };
 
@@ -63,22 +61,7 @@ export const useMenuPopup = (state: UseMenuPopupState) => {
         e.stopPropagation(); // Left and Escape should only close one menu at a time
       }
 
-      // Dismiss is still handled by click, but keydown event should still be propagated
-      // If a child uses keydown without a native click event, the menu will remain open
-      // Only native click event will close
-      if (keyCode === EnterKey || keyCode === SpacebarKey) {
-        dismissedWithKeyboardRef.current = true;
-      }
-
       originalProps?.onKeyDown?.(e);
-    };
-
-    // Assumption made that all clicks will close the popup if propagated from children
-    // To stop clicks from closing the menu call stopPropagation
-    newProps.onClick = (e: React.MouseEvent<HTMLElement>) => {
-      setOpen(e, { open: false, keyboard: dismissedWithKeyboardRef.current });
-      dismissedWithKeyboardRef.current = false;
-      originalProps?.onClick?.(e);
     };
 
     const popupContent = React.createElement(