+++ packages/react-menu/src/utils/isOutsideMenu.ts
@@ -2,7 +2,8 @@ import * as React from 'react';
 import { MenuContextValue } from '../contexts/menuContext';
 
 /**
- * Helper function that checks if a focus event moves focus outside of the menu trigger and menu popup
+ * Helper function that checks if a blur event moves focus outside of the menu trigger and menu popup
+ * Only needed for focus moving between nested submenus with mouse
  */
 export const isOutsideMenu = ({
   triggerRef,
@@ -11,8 +12,14 @@ export const isOutsideMenu = ({
 }: {
   triggerRef: MenuContextValue['triggerRef'];
   menuPopupRef: MenuContextValue['menuPopupRef'];
-  event: React.FocusEvent;
+  event: React.FocusEvent; // onBlur
 }) => {
+  // no related target -> nothing got focus
+  // don't need to handle this because focus did not move between submenus
+  if (!event.relatedTarget) {
+    return false;
+  }
+
   const isOutsidePopup = !menuPopupRef.current?.contains(event.relatedTarget as Node);
   const isOutsideTrigger = !triggerRef.current?.contains(event.relatedTarget as Node);