+++ apps/public-docsite/CHANGELOG.md
@@ -1,8 +1,17 @@
 # Change Log - @fluentui/public-docsite
 
-This log was last generated on Mon, 31 May 2021 07:33:15 GMT and should not be manually modified.
+This log was last generated on Wed, 02 Jun 2021 07:37:15 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [8.2.19](https://github.com/microsoft/fluentui/tree/@fluentui/public-docsite_v8.2.19)
+
+Wed, 02 Jun 2021 07:37:15 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/public-docsite_v8.2.18..@fluentui/public-docsite_v8.2.19)
+
+### Patches
+
+- Bump @fluentui/react-examples to v8.25.1 ([PR #18399](https://github.com/microsoft/fluentui/pull/18399) by bsunderhus@microsoft.com)
 
 ## [8.2.18](https://github.com/microsoft/fluentui/tree/@fluentui/public-docsite_v8.2.18)